Some improvements on reset and dimensions (#2)
* Prevent PHP (7) Error

Solve error with strict standards: Only variables should be passed by reference

* Remove dimensions if any

* Comply with Zend_View_Helper_FormReset

Twitter_Bootstrap3_View_Helper_FormReset must be setup in the same way Zend_View_Helper_FormReset is setup, otherwise this will cause exception.

* Add support for "class" attrib

If adding elements to a form, class is lost. This is especially true for selects and such elements, which do not have a special view helper.
Added manual assignment of class if class is given.